frontend/requirements.txt
numba==0.61.0
numpy==2.0.2
llvmlite==0.44.0
pybind11==2.13.6
pybind11_global==2.13.6
colorama==0.4.6

# required only for manually running tests, bazel has its own absl installation
absl-py==2.1.0
